backend/requirements.txt
beautifulsoup4==4.12.3
lxml==5.1.0
pymongo==4.6.1
python-dotenv==1.0.1
requests==2.31.0
selenium==4.17.2
